Move regex tester to Goodies
