+++ kifi-backend/modules/shoebox/app/com/keepit/admin/AdminRoverController.scala
@@ -122,26 +122,4 @@ class AdminRoverController @Inject() (
     )).map(_ => Redirect(routes.AdminRoverController.getAllUrlRules()))
   }
 
-  def testRegex = AdminUserPage { implicit request =>
-    Ok(views.html.admin.roverTestRegex())
-  }
-
-  def testRegexFilled(regex: String, test: Option[String] = None) = AdminUserPage { implicit request =>
-    Ok(views.html.admin.roverTestRegex(Some(regex), test.map(t => List(t))))
-  }
-
-  def performRegexTest = AdminUserPage(parse.json) { implicit request =>
-    val body = request.body
-    try {
-      val regex = (body \ "regex").as[String]
-      val pattern = Pattern.compile(regex)
-      val tests = (body \ "tests").as[List[String]]
-      val results = tests.map { pattern.matcher(_).matches }
-      Ok(Json.toJson(results))
-    } catch {
-      case e: PatternSyntaxException =>
-        BadRequest(Json.toJson(e.toString))
-    }
-  }
-
 }
