+++ kifi-backend/modules/shoebox/app/views/admin/admin.scala.html
@@ -150,9 +150,15 @@
                 <a class="dropdown-toggle" data-toggle="dropdown" >Rover</a>
                 <ul class="dropdown-menu" role="menu">
                     <li><a href="@com.keepit.controllers.admin.routes.AdminRoverController.searchUrl()">Search for URL</a></li>
-                    <li><a href="@com.keepit.controllers.admin.routes.AdminRoverController.testRegex()">Test Regex</a></li>
                     <li><a href="@com.keepit.controllers.admin.routes.AdminRoverController.getAllProxies()">Proxies</a></li>
                     <li><a href="@com.keepit.controllers.admin.routes.AdminRoverController.getAllUrlRules()">URL Rules</a></li>
+                </ul>
+            </li>
+
+            <li class="dropdown">
+                <a class="dropdown-toggle" data-toggle="dropdown">Goodies</a>
+                <ul class="dropdown-menu" role="menu">
+                    <li><a href="@com.keepit.controllers.admin.routes.AdminGoodiesController.testRegex()">Test Regex</a></li>
                 </ul>
             </li>
         </ul>
