+++ kifi-backend/modules/shoebox/app/views/admin/roverUrlRules.scala.html
@@ -26,7 +26,7 @@
                     <td><input type="text" class="pattern" name="pattern_@urlRule.id.get" value="@urlRule.pattern"></td>
                     <td class="control-group"><input type="text" class="example" name="example_@urlRule.id.get"
                         style="border-color: @if(urlRule.example.matches(urlRule.pattern)) { green } else { red }" value="@urlRule.example"></td>
-                    <td><a href="@com.keepit.controllers.admin.routes.AdminRoverController.testRegexFilled(urlRule.pattern, Some(urlRule.example))" class="btn btn-small">test</a></td>
+                    <td><a href="@com.keepit.controllers.admin.routes.AdminGoodiesController.testRegexFilled(urlRule.pattern, Some(urlRule.example))" class="btn btn-small">test</a></td>
                     <td>
                         <select name="proxy_@urlRule.id.get">
                             <option @if(urlRule.proxy.isEmpty) {selected="selected"} value="">Do not route to proxy</option>
