+++ kifi-backend/modules/shoebox/conf/shoeboxService.routes
@@ -862,9 +862,10 @@ POST          /admin/rover/proxies/create                                       
 GET           /admin/rover/urlRules                                                  @com.keepit.controllers.admin.AdminRoverController.getAllUrlRules()
 POST          /admin/rover/urlRules                                                  @com.keepit.controllers.admin.AdminRoverController.saveUrlRules()
 POST          /admin/rover/urlRules/create                                           @com.keepit.controllers.admin.AdminRoverController.createUrlRule
-GET           /admin/rover/testRegex                                                 @com.keepit.controllers.admin.AdminRoverController.testRegex
-GET           /admin/rover/testRegex/:regex                                          @com.keepit.controllers.admin.AdminRoverController.testRegexFilled(regex: String, test: Option[String])
-POST          /admin/rover/performRegexTest                                          @com.keepit.controllers.admin.AdminRoverController.performRegexTest
+
+GET           /admin/goodies/testRegex                                               @com.keepit.controllers.admin.AdminGoodiesController.testRegex
+GET           /admin/goodies/testRegexFilled                                         @com.keepit.controllers.admin.AdminGoodiesController.testRegexFilled(regex: String, test: Option[String])
+POST          /admin/goodies/performRegexTest                                        @com.keepit.controllers.admin.AdminGoodiesController.performRegexTest
 
 ##########################################
 # Common Healthcheck / service routes
